Clean up test data generation
* Move `cirt-tool-src` to `cmd/util`
* Replace `make certs && make dc` with `make testdata`
* Normalize naming conventions for certificates
* Generate ECH config and key
* Add -host flag for specifying the host name
* Refactor CLI code a bit
