+++ docker-compose.yml
@@ -28,7 +28,7 @@ services:
     stdin_open: true
     tty: true
     volumes:
-      - ./certs:/certs:ro
+      - ./testdata:/testdata:ro
     cap_add:
       - NET_ADMIN
     ulimits:
@@ -53,7 +53,7 @@ services:
     stdin_open: true
     tty: true
     volumes:
-      - ./certs:/certs:ro
+      - ./testdata:/testdata:ro
     cap_add:
       - NET_ADMIN
     ulimits:
@@ -69,7 +69,7 @@ services:
         ipv4_address: 193.167.0.100
         ipv6_address: fd00:cafe:cafe:0::100
     extra_hosts:
-      - "server:193.167.100.100"
+      - "example.com:193.167.100.100"
 
 networks:
   clientnet:
