+++ impl-endpoints/boringssl/run_endpoint.sh
@@ -10,5 +10,5 @@ else
     echo "Running BoringSSL server."
     echo "Server params: $SERVER_PARAMS"
     echo "Test case: $TESTCASE"
-    bssl server -accept 4433 -cert /certs/server.cert -key /certs/server.key -subcert /certs/dc.txt
+    bssl server -accept 4433 -cert /testdata/example.crt -key /testdata/example.key -subcert /testdata/dc.txt
 fi
