+++ impl-endpoints/rustls/run_endpoint.sh
@@ -12,5 +12,5 @@ else
     echo "Running rustls server."
     echo "Server params: $SERVER_PARAMS"
     echo "Test case: $TESTCASE"
-    tlsserver --certs /certs/server.cert --key /certs/server.key -p 4433 echo
+    tlsserver --certs /testdata/example.crt --key /testdata/example.key -p 4433 echo
 fi
